Update Icon stories
